pyproject.toml
[project]
name = "repl-nix-textflownavigator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "markdown>=3.7",
    "replit>=4.1.0",
    "flask-cors>=5.0.0",
    "replit-object-storage==1.0.2",
    "werkzeug",
    "pymupdf==1.23.7",
    "python-dotenv>=1.0.1",
    "openai>=1.54.4",
    "pypdf2>=3.0.1",
    "reportlab>=4.2.5",
    "pillow>=11.0.0",
    "asgiref>=3.8.1",
    "hypercorn>=0.17.3",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "tenacity>=9.0.0",
    "requests>=2.32.3",
    "oauthlib>=3.2.2",
    "pyopenssl>=24.2.1",
    "flask-sock>=0.7.0",
    "fitz>=0.0.1.dev2",
    "tools",
]
